backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
chromadb==0.5.20
openai==1.54.0
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
pydantic==2.10.0
python-multipart==0.0.12
